[physical-plan]: rename aggregate stream modules
