+++ datafusion/physical-plan/src/aggregates/partial_reduce_stream.rs
@@ -18,7 +18,7 @@
 //! Partial-reduce hash aggregation stream implementation.
 //!
 //! This stream is part of the incremental migration from
-//! [`crate::aggregates::row_hash::GroupedHashAggregateStream`].
+//! [`crate::aggregates::grouped_hash_stream::GroupedHashAggregateStream`].
 //!
 //! See issue for details: <https://github.com/apache/datafusion/issues/22710>
