+++ datafusion/physical-plan/src/aggregates/skip_partial.rs
@@ -22,7 +22,7 @@ use crate::metrics;
 /// Tracks if the aggregate should skip partial aggregations
 ///
 /// See "partial aggregation" discussion on
-/// [`crate::aggregates::row_hash::GroupedHashAggregateStream`].
+/// [`crate::aggregates::grouped_hash_stream::GroupedHashAggregateStream`].
 pub(super) struct SkipAggregationProbe {
     // ========================================================================
     // PROPERTIES:
@@ -117,7 +117,7 @@ impl SkipAggregationProbe {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::aggregates::row_hash::GroupedHashAggregateStream;
+    use crate::aggregates::grouped_hash_stream::GroupedHashAggregateStream;
     use crate::aggregates::{AggregateExec, AggregateMode, PhysicalGroupBy};
     use crate::execution_plan::ExecutionPlan;
     use crate::test::TestMemoryExec;
@@ -134,7 +134,7 @@ mod tests {
     use datafusion_physical_expr::expressions::col;
     use futures::StreamExt;
 
-    // Migrated to PartialHashAggregateStream coverage in hash_aggregate.rs;
+    // Migrated to PartialHashAggregateStream coverage in hash_stream.rs;
     // kept here for the legacy GroupedHashAggregateStream implementation.
     #[tokio::test]
     async fn test_skip_aggregation_probe_not_locked_until_skip() -> Result<()> {
